Neural Network Programme
Using test data included with the tensorflow keras import, this programme can predict what item of clothing something is with an accuracy of around 85%.
Use the predict_this() function to input a number. It will show you a graph of the item, and the guess will be printed to console.