Guess The Number
Programme intended for the user to enter 2 integers, then decide if they are sure about this, then guess a random number between these two integers using clues such as 'higher' and 'lower'.
Particular attention was paid towards user input. If the input could not be converted to an int, then a print statement asking for integers is put out, and the original function is called again.