Arithmetic Formatter
Programme that takes up to 5 string arguments within a list, such that the string arguments are of form 'x +/- y'. There is an optional argument of 'True' if you would like to see the answer with the formatted equation.
The maximum is 5 arguments. The maximum is 4 digits per number. The only accepted operations are + and -.  Only digits are accepted as the numbers.

This was developed for a freeCodeCamp project, however I still haven't managed to get the formed up equations next to each other, only one after another. I will revisit this when I have more knowledge of how to do this.